# docs/api/openapi-query.yaml
openapi: 3.1.0
info:
  title: siglake query
  description: |-
    DataFusion SQL over the Iceberg tables, index and index-template management, batch jobs, GDPR delete tasks, and a Jaeger-compatible trace read API.

    Ingest lives on a separate server (`openapi-ingest.yaml`) — this one only reads, apart from index management and delete tasks.
  license:
    name: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8089
  description: Default `--bind` address
paths:
  /api/v1/delete-tasks:
    get:
      tags:
      - delete-tasks
      summary: List delete tasks and their progress.
      operationId: list
      parameters:
      - name: index_id
        in: query
        description: Return only tasks for this index. Omit to list every index's tasks.
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Matching delete tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeleteTask'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
    post:
      tags:
      - delete-tasks
      summary: Submit a delete task (GDPR / targeted deletion).
      description: |-
        Asynchronous: this records the request and returns immediately with the task
        in `pending`. A background executor rewrites the affected data files without
        their matching rows. Rows removed from the live snapshot stay visible to
        snapshot time-travel until a later snapshot-expiry plus orphan-GC sweep drops
        the old files, so completion here is not the same as unrecoverable erasure.

        THE TASK BELONGS TO ONE INDEX INCARNATION. The record carries the
        `table_uuid` this server resolved `index_id` to at submission. An index id
        is reusable — deleting an index and creating it again under the same id
        gives a different table — and the executor refuses a task whose `index_id`
        no longer resolves to that table, so a deletion queued against the dropped
        index never touches the replacement's rows. The refusal is a `failed` task
        whose `error` names both tables; recover it the same way as any other
        failure, by resubmitting against the index that exists. Tasks recorded
        before this binding existed carry a null `table_uuid` and are refused for
        the same reason: nothing infers an incarnation from the reused name.

        RECOVERING A `failed` TASK. The lifecycle is one-way: the executor takes
        `pending` tasks only, and no surface returns a terminal task to `pending`.
        A task that reached `failed` therefore never runs again, even though the
        deletion request was acknowledged with a 201 — and neither does one left in
        `running` by a crash or by a status write that failed after its rewrite
        committed. The recovery is the same for both: remedy the failure cause —
        named in the task's `error`, where it has one — make sure the executor runs
        the fixed build, and then POST the original `index_id`, `predicate_sql`,
        `start_ts` and `end_ts` here again under the same tenant identity — request
        fields only. The response carries a NEW `task_id` in `pending`; the failed
        record and its `error` are left exactly as they were, so record both ids in
        your audit trail. Then confirm delete-task execution was not turned off for
        this deployment — the compactor's maintenance sweep runs it by default,
        `SIGLAKE_DELETE_TASKS=0` disables it — and poll
        `GET /api/v1/delete-tasks/{id}` on the new id until it is `done` or
        `failed`.

        This is a resubmission, not a retry: it is not idempotent at the HTTP layer
        (every POST creates another task) and it promises nothing about
        exactly-once execution. Each task reports only what its own run rewrote,
        never the request's cumulative effect. Over an unchanged dataset with the
        same predicate and the same fixed bounds a repeat finishes with
        `rows_deleted: 0`, `files_rewritten: 0` and no new snapshot — but a
        predicate that is time-dependent, or new rows that have arrived since,
        makes the repeat delete a different set.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeleteTaskRequest'
        required: true
      responses:
        '201':
          description: Task recorded in `pending`, bound to the index incarnation named in `table_uuid`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteTask'
        '400':
          description: Empty or unparseable `predicate_sql`, or `start_ts >= end_ts`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such index.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
  /api/v1/delete-tasks/{id}:
    get:
      tags:
      - delete-tasks
      summary: Fetch one delete task, including rows deleted and files rewritten.
      operationId: get
      parameters:
      - name: id
        in: path
        description: Task id.
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: The delete task. Pending tasks include a read-only claim observation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteTaskGetResponse'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
...
components:
  schemas:
    ApiErrorBody:
      type: object
      description: |-
        Error body used by every endpoint on this server.

        Spec-only. Some errors additionally flatten diagnostic keys into the top
        level — `cost` on a rejected or timed-out query, and `cost` plus `stats` on
        the mid-flight rows breaker — so this schema deliberately allows additional
        properties rather than pinning an exact shape.
      required:
      - error
      - code
      properties:
        code:
          type: integer
          format: int32
          description: Echo of the HTTP status code.
          minimum: 0
        cost:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CostReport'
            description: Present when the failure was cost- or time-related.
        error:
          type: string
          description: Human-readable failure description.
...
    ComplexityClass:
      type: string
      enum:
      - small
      - medium
      - large
      - huge
    CostReport:
      type: object
      description: |-
        Cost report returned by `/explain` and attached to every successful
        query response.
      required:
      - estimated_bytes_scanned
      - estimated_rows_processed
      - estimated_runtime_seconds
      - complexity_class
      - warnings
      - exact
      properties:
        complexity_class:
          $ref: '#/components/schemas/ComplexityClass'
        estimated_bytes_scanned:
          type: integer
          format: int64
          minimum: 0
        estimated_rows_processed:
          type: integer
          format: int64
          minimum: 0
        estimated_runtime_seconds:
          type: number
          format: double
        exact:
          type: boolean
          description: |-
            `true` when the underlying statistics are exact; `false` when
            they fell back to a heuristic.
        files_considered:
          type:
          - integer
          - 'null'
          description: |-
            Live files considered BEFORE the manifest time-bounds prune —
            `files_considered - files_to_scan` is planning-time pruning's win.
          minimum: 0
        files_to_scan:
          type:
          - integer
          - 'null'
          minimum: 0
        warnings:
          type: array
          items:
            type: string
    CreateDeleteTaskRequest:
      type: object
      required:
      - index_id
      - predicate_sql
      properties:
        end_ts:
          type:
          - string
          - 'null'
          format: date-time
          description: Optional exclusive upper time bound.
        index_id:
          type: string
          description: Index whose rows should be deleted.
        predicate_sql:
          type: string
          description: |-
            SQL boolean expression selecting the rows to delete, as it would appear
            after `WHERE`.
          example: host = 'web-01'
        start_ts:
          type:
          - string
          - 'null'
          format: date-time
          description: |-
            Optional inclusive lower time bound, narrowing the files the executor
            has to rewrite.
    DeleteTask:
      type: object
      description: |-
        A GDPR/delete-task: the deletion request, its lifecycle state, and — once
        terminal — how much it rewrote. Immutable apart from state transitions and
        that terminal accounting.
      required:
      - task_id
      - index_id
      - predicate_sql
      - created_at
      - state
      - files_rewritten
      - rows_deleted
      properties:
        created_at:
          type: string
          format: date-time
        end_ts:
          type:
          - string
          - 'null'
          format: date-time
        error:
          type:
          - string
          - 'null'
        files_rewritten:
          type: integer
          minimum: 0
        index_id:
          type: string
        predicate_sql:
          type: string
        rows_deleted:
          type: integer
          format: int64
          minimum: 0
        start_ts:
          type:
          - string
          - 'null'
          format: date-time
        state:
          $ref: '#/components/schemas/DeleteTaskState'
        table_uuid:
          type:
          - string
          - 'null'
          description: |-
            The Iceberg table this deletion was accepted against — the index
            incarnation it is authorised for. An index id can be deleted and
            created again, and the executor refuses a task whose index id now
            resolves to a different table, so the request recovers by resubmission
            instead of deleting rows nobody reviewed.

            Null on tasks recorded by builds before this binding existed. Those are
            readable, and refused at execution for the same reason.
        task_id:
          type: string
          format: uuid
    DeleteTaskClaimObservation:
      type: object
      description: |-
        Read-only observation of a pending delete task's permanent claim object.

        `present` says only that the object was observed during this read. It does
        not establish claimant liveness, abandonment, execution, commit outcome,
        or whether takeover would be safe. When a present claim body is malformed
        or names another task, its diagnostic fields are null rather than treating
        the claim as absent.
      required:
      - present
      - observed_at
      properties:
        age_seconds:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Whole seconds from claim creation to `observed_at`, floored at zero for
            clock skew. This is not execution duration or proof of abandonment.
          minimum: 0
        claimant:
          type:
          - string
          - 'null'
          format: uuid
        claimed_at:
          type:
          - string
          - 'null'
          format: date-time
        observed_at:
          type: string
          format: date-time
        present:
          type: boolean
    DeleteTaskGetResponse:
      allOf:
      - $ref: '#/components/schemas/DeleteTask'
      - type: object
        properties:
          claim:
            oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/DeleteTaskClaimObservation'
              description: Present only while the persisted task state is `pending`.
      description: |-
        GET-only representation of a delete task. POST and collection-list
        responses remain the persisted [`DeleteTask`] contract.
    DeleteTaskState:
      type: string
      description: |-
        Delete-task lifecycle state. The ledger is immutable apart from state
        transitions and terminal stats/error population.
      enum:
      - pending
      - running
      - done
      - failed
...
  securitySchemes:
    query_token:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <token>`. With OIDC configured the token is a JWT and its claims select the tenant; otherwise it is matched against the cluster''s static token list. A server started with neither accepts unauthenticated requests.'
security:
- query_token: []
...